academia/requirements.txt
fastapi
pydantic
uvicorn
peewee
mysqlclient
python-dotenv
grpcio
grpcio-tools
pyjwt